fixed a mistake where i compared size_t with int

+++ libft/src/mem/ft_realloc_str_arr.c
@@ -6,7 +6,7 @@
 /*   By: tblaase <tblaase@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/28 20:35:01 by tblaase           #+#    #+#             */
-/*   Updated: 2021/10/28 21:05:47 by tblaase          ###   ########.fr       */
+/*   Updated: 2021/10/28 21:11:35 by tblaase          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ new: adress of the reallocated array of strings
 	int		i;
 
 	i = 0;
-	if (size <= ft_str_arr_len(arr))
+	if (size <= (size_t)ft_str_arr_len(arr))
 		return (arr);
 	new = ft_calloc(size, sizeof(char *));
 	if (!new)
